+++ fixtures/core_test/a.clj
@@ -6,7 +6,3 @@
     [z :as-alias z])
   (:import
     [java.io File]))
-
-::b/kw
-::z/kw
-File